Update employee update details screen and form, refactored into three widgets
