+++ lib/employee/utilities/dashboard_item_list.dart
@@ -1,7 +1,7 @@
 import 'package:echno_attendance/employee/hr_screens/add_employee.dart';
 import 'package:echno_attendance/employee/hr_screens/attendance_report/employee_attendancereport.dart';
 import 'package:echno_attendance/employee/hr_screens/employee_register.dart';
-import 'package:echno_attendance/employee/hr_screens/update_details.dart';
+import 'package:echno_attendance/employee/hr_screens/update_details_screen.dart';
 import 'package:echno_attendance/employee/models/dashboard_item.dart';
 import 'package:echno_attendance/leave_module/screens/leave_register.dart';
 import 'package:flutter/material.dart';
@@ -24,7 +24,7 @@ class DashboardItemList {
         text: 'Update Details',
         onTap: () {
           Navigator.push(context, MaterialPageRoute(builder: (context) {
-            return const UpdateEmployeeDetails();
+            return const UpdateEmployeeDetailsScreen();
           }));
         },
       ),
